refactor(release): label-mastermind now extracted from here
