+++ packages/release-mastermind/src/conditions/util/creatorMatches.ts
@@ -12,7 +12,7 @@ export const creatorMatches = (
   condition: ConditionCreatorMatches,
   issue: IssueProps | PRProps
 ) => {
-  const pattern = utils.processRegExpPattern(condition.pattern)
+  const pattern = utils.parsingData.processRegExpPattern(condition.pattern)
   return pattern.test(issue.creator)
 }
