+++ packages/release-mastermind/src/conditions/util/descriptionMatches.ts
@@ -12,7 +12,7 @@ const descriptionMatches = (
   condition: ConditionDescriptionMatches,
   issue: IssueProps | PRProps
 ) => {
-  const pattern = utils.processRegExpPattern(condition.pattern)
+  const pattern = utils.parsingData.processRegExpPattern(condition.pattern)
   return pattern.test(issue.description)
 }
