+++ packages/release-mastermind/src/utils/helper/semantic.ts
@@ -1,0 +1,21 @@
+export const semantic: string[] = [
+  'chore',
+  'opt',
+  'optimisation',
+  'style',
+  'maint',
+  'maintenance',
+  'ref',
+  'refactor',
+  'revert',
+  'dep',
+  'deprecated',
+  'removal',
+  'docs',
+  'documentation',
+  'feat',
+  'enhance',
+  'feature',
+  'enhancement',
+  'fix'
+]
